+++ CHANGELOG.md
@@ -36,6 +36,12 @@
     * deprecated `stubbles\lang\ResourceLoader::getResourceUris()´, use `stubbles\lang\ResourceLoader::availableResourceUris()` instead
     * deprecated `stubbles\lang\ResourceLoader::getRootPath()´ and `stubbles\lang\ResourceLoader::getRoot()´, use `stubbles\lang\Rootpath` instead
     * deprecated `stubbles\peer\http\AcceptHeader::getList()`
+    * deprecated `stubbles\peer\BsdSocket::getType()`, use `stubbles\peer\BsdSocket::type()` instead
+    * deprecated `stubbles\peer\BsdSocket::getOption()`, use `stubbles\peer\BsdSocket::option()` instead
+    * deprecated `stubbles\peer\Socket::getTimeout()`, use `stubbles\peer\Socket::timeout()` instead
+    * deprecated `stubbles\peer\Socket::getPrefix()`, use `stubbles\peer\Socket::usesSsl()` instead
+    * deprecated `stubbles\peer\Socket::getInputStream()`, use `stubbles\peer\Socket::in()` instead
+    * deprecated `stubbles\peer\Socket::getOutputStream()`, use `stubbles\peer\Socket::out()` instead
     * deprecated `stubbles\peer\Uri::getScheme()`, use `stubbles\peer\Uri::scheme()` instead
     * deprecated `stubbles\peer\Uri::getUser()`, use `stubbles\peer\Uri::user()` instead
     * deprecated `stubbles\peer\Uri::getPassword()`, use `stubbles\peer\Uri::password()` instead
@@ -58,6 +64,10 @@
     * deprecated `stubbles\peer\http\HttpResponse::getReasonPhrase()`, use `stubbles\peer\http\HttpResponse::reasonPhrase()` instead
     * deprecated `stubbles\peer\http\HttpResponse::getHeader()`, use `stubbles\peer\http\HttpResponse::headers()` instead
     * deprecated `stubbles\peer\http\HttpResponse::getBody()`, use `stubbles\peer\http\HttpResponse::body()` instead
+  * deprecated `stubbles\peer\BsdSocket::getDomain()`, will be removed with 5.0.0
+  * deprecated `stubbles\peer\Socket::getHost()`, will be removed with 5.0.0
+  * deprecated `stubbles\peer\Socket::getPort()`, will be removed with 5.0.0
+  * `stubbles\peer\Socket::getInputStream()` and `stubbles\peer\Socket::getOutputStream()` will now always return the same instance
   * `stubbles\peer\http\HttpUri::fromString()` no longer accepts uris with userinfo by default in compliance with RFC 7230, to retain the old behaviour pass `stubbles\peer\http\Http::RFC_2616` as second parameter